+++ test/xcop.test.js
@@ -20,6 +20,7 @@ const PACKS = [
   ...allFilesFrom(path.resolve(__dirname, 'resources', 'xpath-format-packs')),
   ...allFilesFrom(path.resolve(__dirname, 'resources', 'axis-packs')),
   ...allFilesFrom(path.resolve(__dirname, 'resources', 'namespace-packs')),
+  ...allFilesFrom(path.resolve(__dirname, 'resources', 'result-namespace-packs')),
   ...allFilesFrom(path.resolve(__dirname, 'resources', 'node-set-packs')),
   ...allFilesFrom(path.resolve(__dirname, 'resources', 'count-packs')),
   ...allFilesFrom(path.resolve(__dirname, 'resources', 'string-length-packs')),